% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ALIGN_CDR3S}
\alias{ALIGN_CDR3S}
\title{Whether to align CDR3 sequences before distance computation}
\format{
A length-1 logical.
}
\usage{
ALIGN_CDR3S
}
\description{
Logical flag controlling CDR3 alignment in TCR distance calculations.
}
\keyword{internal}
