% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\docType{data}
\name{BASE_PARTNER}
\alias{BASE_PARTNER}
\title{Nucleotide complement mapping}
\format{
A named character vector of length 17.
}
\usage{
BASE_PARTNER
}
\description{
Named character vector mapping each nucleotide (and IUPAC ambiguity code and
gap character) to its Watson-Crick complement. Matches the Python
\code{base_partner} dictionary used by CoNGA.
}
\keyword{internal}
