% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_generic.R
\name{.detect_format_name}
\alias{.detect_format_name}
\title{Identify the format name without transforming the data.frame}
\usage{
.detect_format_name(df)
}
\description{
Identify the format name without transforming the data.frame
}
\keyword{internal}
