% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcr_clumping.R
\name{.empty_clumping_results}
\alias{.empty_clumping_results}
\title{Create an empty clumping results data.frame}
\usage{
.empty_clumping_results()
}
\value{
A zero-row data.frame with the correct columns.
}
\description{
Create an empty clumping results data.frame
}
\keyword{internal}
