% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcr_db_matching.R
\name{.empty_tcr_db_match_df}
\alias{.empty_tcr_db_match_df}
\title{Empty result data.frame for TCR database matching}
\usage{
.empty_tcr_db_match_df()
}
\description{
Empty result data.frame for TCR database matching
}
\keyword{internal}
