% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_distance.R
\name{.zero_v_dist_matrix}
\alias{.zero_v_dist_matrix}
\title{Create a zero V-region distance matrix with matching gene names}
\usage{
.zero_v_dist_matrix(v_dist)
}
\arguments{
\item{v_dist}{A named square \code{NumericMatrix} from
\code{.compute_v_region_distance_matrix()}.}
}
\value{
A zero matrix with the same dimensions and row/column names.
}
\description{
Create a zero V-region distance matrix with matching gene names
}
\keyword{internal}
