% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GAP_CHARACTER}
\alias{GAP_CHARACTER}
\title{Gap character used in sequence alignments}
\format{
A length-1 character string.
}
\usage{
GAP_CHARACTER
}
\description{
Single character representing a gap in aligned sequences.
}
\keyword{internal}
