% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GAP_PENALTY_CDR3_REGION}
\alias{GAP_PENALTY_CDR3_REGION}
\title{Gap penalty for CDR3-region alignments}
\format{
A length-1 integer.
}
\usage{
GAP_PENALTY_CDR3_REGION
}
\description{
Integer gap penalty applied when aligning CDR3-region sequences.
}
\keyword{internal}
