% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GAP_PENALTY_V_REGION}
\alias{GAP_PENALTY_V_REGION}
\title{Gap penalty for V-region alignments}
\format{
A length-1 integer.
}
\usage{
GAP_PENALTY_V_REGION
}
\description{
Integer gap penalty applied when aligning V-region sequences.
}
\keyword{internal}
