% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_genes.R
\name{list_organisms}
\alias{list_organisms}
\title{List available organisms in the gene database}
\usage{
list_organisms()
}
\value{
A character vector of organism names, sorted alphabetically.
}
\description{
Returns the names of all organisms available in the bundled gene
database. These are the valid values for the \code{organism} parameter
in functions like \code{\link{tcrdist_matrix}} and \code{\link{TCRrep}}.
}
\examples{
list_organisms()

}
\seealso{
\code{\link{load_gene_database}}
}
