% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_blosum62_lookup}
\alias{rcpp_blosum62_lookup}
\title{Look up a single BLOSUM62 score for two amino acids}
\usage{
rcpp_blosum62_lookup(aa1, aa2)
}
\arguments{
\item{aa1}{Single-letter amino acid code (length-1 character string).}

\item{aa2}{Single-letter amino acid code (length-1 character string).}
}
\value{
Integer BLOSUM62 score for the pair.
}
\description{
Look up a single BLOSUM62 score for two amino acids
}
\keyword{internal}
