% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code.R
\docType{data}
\name{REVERSE_GENETIC_CODE}
\alias{REVERSE_GENETIC_CODE}
\title{Reverse genetic code (amino acid to codons)}
\format{
A named list of length 21 (20 amino acids + stop).
}
\usage{
REVERSE_GENETIC_CODE
}
\description{
Named list mapping each single-letter amino acid code (and \code{"*"} for
stop codons) to a character vector of the corresponding lowercase DNA
codons.
}
\keyword{internal}
