% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{subset,TCRrep-method}
\alias{subset,TCRrep-method}
\title{Subset a TCRrep using non-standard evaluation}
\usage{
\S4method{subset}{TCRrep}(x, subset, ...)
}
\arguments{
\item{x}{A \code{TCRrep} object.}

\item{subset}{A logical expression evaluated in the context of
\code{x@clone_df}. Column names can be used directly.}

\item{...}{Ignored.}
}
\value{
A new \code{TCRrep} object with only matching clonotypes.
}
\description{
Filter clonotypes using column expressions evaluated against
\code{clone_df}. This is the recommended way to filter a TCRrep:
}
\details{
\preformatted{rep |> subset(epitope == "PA")}
}
\examples{
\donttest{
data(dash)
rep <- TCRrep(dash, organism = "mouse", compute_distances = TRUE)
pa <- subset(rep, epitope == "PA")
pa
}

}
