% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{subset-TCRrep}
\alias{subset-TCRrep}
\alias{[,TCRrep,ANY,ANY,ANY-method}
\title{Subset a TCRrep object by row indices}
\usage{
\S4method{[}{TCRrep,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{TCRrep} object.}

\item{i}{Row indices: logical, integer, or character (rownames).}

\item{j}{Ignored.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\value{
A new \code{TCRrep} object with the selected clonotypes.
}
\description{
Subsets a \code{\link{TCRrep}} by clonotype indices. Filters
\code{clone_df} and all associated distance matrices simultaneously.
KNN matrices and meta-clonotypes are cleared since they reference the
original indexing.
}
\examples{
\donttest{
data(dash)
rep <- TCRrep(dash, organism = "mouse", compute_distances = TRUE)
pa <- rep[rep@clone_df$epitope == "PA", ]
pa
}

}
