% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{tcr_gini}
\alias{tcr_gini}
\title{Gini index of a clonotype distribution}
\usage{
tcr_gini(counts)
}
\arguments{
\item{counts}{Integer vector. Clonotype counts (positive integers).}
}
\value{
A numeric scalar between 0 and 1.
}
\description{
Computes the Gini coefficient measuring inequality in clonotype
abundances. Values near 0 indicate a uniform (equal) distribution;
values near 1 indicate extreme inequality (one dominant clone).
}
\examples{
# Uniform: Gini near 0
tcr_gini(rep(10, 5))

# Dominated: Gini near 1
tcr_gini(c(10000, 1, 1))

}
\seealso{
\code{\link{tcr_shannon_entropy}}, \code{\link{tcr_clonality}},
\code{\link{tcr_diversity}}
}
