% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{tcr_richness}
\alias{tcr_richness}
\title{TCR repertoire richness}
\usage{
tcr_richness(counts)
}
\arguments{
\item{counts}{Integer vector. Clonotype counts (positive integers).}
}
\value{
An integer: the number of unique clonotypes.
}
\description{
Returns the number of unique clonotypes (species richness).
}
\examples{
tcr_richness(c(10, 5, 3, 1, 1))  # 5

}
\seealso{
\code{\link{tcr_diversity}}, \code{\link{tcr_clonality}}
}
