% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrdistR-package.R
\docType{package}
\name{tcrdistR-package}
\alias{tcrdistR}
\alias{tcrdistR-package}
\title{tcrdistR: TCR Distance Calculations for T-Cell Receptor Repertoire Analysis}
\description{
C++-accelerated TCR distance calculations for T-cell receptor repertoire analysis. Computes pairwise TCRdist distances incorporating V-region and CDR3 sequence comparisons using BLOSUM62-derived substitution matrices. Targets feature parity with Python tcrdist3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shihanli92/tcrdistR}
  \item Report bugs at \url{https://github.com/shihanli92/tcrdistR/issues}
}

}
\author{
\strong{Maintainer}: Shihan Li \email{shihanli1992@gmail.com}

}
\keyword{internal}
