% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TRIM_CDR3S}
\alias{TRIM_CDR3S}
\title{Whether to trim CDR3 sequences before distance computation}
\format{
A length-1 logical.
}
\usage{
TRIM_CDR3S
}
\description{
Logical flag controlling CDR3 trimming in TCR distance calculations.
}
\keyword{internal}
