% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_CDR3_REGION}
\alias{WEIGHT_CDR3_REGION}
\title{Weight for CDR3-region distance contribution}
\format{
A length-1 integer.
}
\usage{
WEIGHT_CDR3_REGION
}
\description{
Integer weight applied to CDR3-region distances when computing combined TCR
distances.
}
\keyword{internal}
