% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WEIGHT_V_REGION}
\alias{WEIGHT_V_REGION}
\title{Weight for V-region distance contribution}
\format{
A length-1 integer.
}
\usage{
WEIGHT_V_REGION
}
\description{
Integer weight applied to V-region distances when computing combined TCR
distances.
}
\keyword{internal}
