% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AMINO_ACIDS}
\alias{AMINO_ACIDS}
\title{Standard 20 amino acid single-letter codes}
\format{
A character vector of length 20.
}
\usage{
AMINO_ACIDS
}
\description{
A character vector of the 20 standard amino acids in alphabetical order,
ported from Python CoNGA \code{amino_acids.py}.
}
\keyword{datasets}
