% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_distance.R
\name{hamming_matrix}
\alias{hamming_matrix}
\title{Pairwise Hamming distance matrix for CDR3 sequences}
\usage{
hamming_matrix(cdr3_seqs)
}
\arguments{
\item{cdr3_seqs}{Character vector. CDR3 amino acid sequences.}
}
\value{
An integer matrix of dimensions N x N.
}
\description{
Computes an N x N integer matrix of Hamming distances between CDR3
sequences. For pairs of equal length, the distance is the number of
mismatches. For pairs of unequal length, the distance is set to the
length of the longer sequence (maximum penalty).
}
\examples{
seqs <- c("CASSI", "CASSK", "CASRL")
hamming_matrix(seqs)
}
\seealso{
\code{\link{hamming_distance}}, \code{\link{tcrdist_matrix}}
}
