% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distances.R
\name{plot_distance_distribution}
\alias{plot_distance_distribution}
\title{Plot the distribution of pairwise TCRdist distances}
\usage{
plot_distance_distribution(dist_matrix, title = NULL, binwidth = NULL)
}
\arguments{
\item{dist_matrix}{Numeric matrix. Square pairwise distance matrix.}

\item{title}{Optional plot title.}

\item{binwidth}{Numeric. Histogram bin width. If \code{NULL}, uses
ggplot2 default.}
}
\value{
A \code{ggplot} object.
}
\description{
Histogram with overlaid density curve of the upper triangle of a pairwise
distance matrix. A vertical dashed line marks the median distance.
}
\examples{
\dontrun{
mat <- matrix(c(0,10,20, 10,0,15, 20,15,0), nrow = 3)
plot_distance_distribution(mat)
}

}
\seealso{
\code{\link{plot_tcrdist_heatmap}}, \code{\link{tcrdist_matrix}}
}
